bit_errors: make cell address uint32_t like spec says Chip addresses can be more than 16 bits, so this has to be larger

+++ source/bit_errors.h
@@ -18,7 +18,7 @@
 typedef struct MDE_Error_Data MDE_Error_Data_t;
 struct MDE_Error_Data {
     uint8_t chip_id;             // Chip number
-    uint16_t cell_address;       // Byte address
+    uint32_t cell_address;       // Byte address
     uint8_t written_sequence;    // Written sequence (all 0s or all 1s)
     uint8_t retrieved_sequence;  // Read sequence
     MDE_Error_Data_t *next; // Pointer to the next link.
